Refactor: Spot 네비게이션
